Move bulletin board to new page and add image carousel
