+++ src/components/Board/BoardPage.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import Spotlight from '../Home/Spotlight/Index';
+
+function BoardPage() {
+  return <Spotlight />;
+}
+
+export default BoardPage;
